# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Create a new user account and atomically bootstrap a personal workspace, admin membership, workspace settings, and free workspace subscription. A verification OTP will be sent to the provided email.",
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully, verification email sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already exists"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Register new user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate with email and password. User must have verified email and at least one active workspace membership. Returns workspace-aware access and refresh tokens. Web clients receive refresh token as HttpOnly cookie.",
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Account not verified or suspended"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Login user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "description": "Verify email address using the 6-digit OTP sent to email. On success, returns workspace-aware tokens for auto-login. Web clients receive refresh token as HttpOnly cookie.",
        "operationId": "AuthController_verifyEmail_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully, user logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP code"
          },
          "429": {
            "description": "Too many failed attempts"
          }
        },
        "summary": "Verify email with OTP",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "description": "Request a new verification OTP. Rate limited to prevent abuse (60s cooldown, 5/hour max).",
        "operationId": "AuthController_resendVerification_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email sent (if account exists)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Email already verified"
          },
          "429": {
            "description": "Rate limit exceeded or cooldown period active"
          }
        },
        "summary": "Resend verification email",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "description": "Request a password reset OTP. Always returns success to prevent email enumeration.",
        "operationId": "AuthController_forgotPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent (if account exists)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Request password reset",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "description": "Reset password using the 6-digit OTP. The OTP is verified and consumed in a single step. All active sessions will be terminated.",
        "operationId": "AuthController_resetPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired reset code, or password validation failed"
          },
          "429": {
            "description": "Too many failed attempts"
          }
        },
        "summary": "Reset password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Get new access and refresh tokens while preserving active workspace context. The used refresh token is revoked (rotation). Web clients send refresh token via HttpOnly cookie; mobile clients send it in the request body.",
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Refresh tokens",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revoke the refresh token to logout from current device. Web clients send refresh token via HttpOnly cookie; mobile clients send it in the request body.",
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "description": "Revoke all refresh tokens to logout from all devices.",
        "operationId": "AuthController_logoutAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out from all devices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout from all devices",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Get the profile of the currently authenticated user together with active workspace and membership context.",
        "operationId": "AuthController_getMe_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/me/avatar": {
      "put": {
        "description": "Upload a new avatar for the authenticated user. Replaces the previous avatar if one already exists.",
        "operationId": "AuthController_uploadAvatar_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadAvatarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          },
          "422": {
            "description": "Invalid file type or file too large"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload or replace avatar",
        "tags": [
          "Authentication"
        ]
      },
      "delete": {
        "description": "Remove the current avatar for the authenticated user. Safe to call even when the user has no avatar.",
        "operationId": "AuthController_removeAvatar_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Avatar removed successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove avatar",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Facets API",
    "description": "Professional Finance Tracker SaaS API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password (min 8 chars, must contain uppercase, lowercase, and number)",
            "example": "SecurePass123",
            "minLength": 8,
            "maxLength": 128
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
...
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Registration successful. Please check your email to verify your account."
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "message",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "SecurePass123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TokensResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token for API authentication",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token for obtaining new access tokens",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "expiresIn": {
            "type": "number",
            "description": "Access token expiration time in seconds",
            "example": 3600
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/TokensResponseDto"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "tokens",
          "user"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "description": "6-digit OTP code received via email",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "VerifyEmailResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Email verified successfully"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokensResponseDto"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "message",
          "tokens",
          "user"
        ]
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "description": "6-digit OTP code received via email",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          },
          "newPassword": {
            "type": "string",
            "description": "New password (min 8 chars, must include uppercase, lowercase, and number)",
            "example": "NewPassword123",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "code",
          "newPassword"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token received during login or previous refresh. Required for mobile clients. Web clients can omit this and use the HttpOnly cookie instead.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "UploadAvatarDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "Avatar image file (JPEG, PNG, or WebP, max 2MB)."
          }
        },
        "required": [
          "file"
        ]
      },
...
    }
  }
}